refactored packages to organzie code better and added an example
