--- javav2/example_code/cloudwatch/src/main/java/com/example/cloudwatch/DeleteSubscriptionFilter.java
@@ -1,63 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.example.cloudwatch;
-
-// snippet-start:[cloudwatch.java2.delete_subscription_filter.main]
-// snippet-start:[cloudwatch.java2.delete_subscription_filter.import]
-import software.amazon.awssdk.services.cloudwatch.model.CloudWatchException;
-import software.amazon.awssdk.services.cloudwatchlogs.CloudWatchLogsClient;
-import software.amazon.awssdk.services.cloudwatchlogs.model.DeleteSubscriptionFilterRequest;
-// snippet-end:[cloudwatch.java2.delete_subscription_filter.import]
-
-/**
- * Before running this Java V2 code example, set up your development
- * environment, including your credentials.
- *
- * For more information, see the following documentation topic:
- *
- * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
- */
-public class DeleteSubscriptionFilter {
-    public static void main(String[] args) {
-        final String usage = """
-
-                Usage:
-                  <filter> <logGroup>
-
-                Where:
-                  filter - The name of the subscription filter (for example, MyFilter).
-                  logGroup - The name of the log group. (for example, testgroup).
-                """;
-
-        if (args.length != 2) {
-            System.out.println(usage);
-            System.exit(1);
-        }
-
-        String filter = args[0];
-        String logGroup = args[1];
-        CloudWatchLogsClient logs = CloudWatchLogsClient.builder()
-                .build();
-
-        deleteSubFilter(logs, filter, logGroup);
-        logs.close();
-    }
-
-    public static void deleteSubFilter(CloudWatchLogsClient logs, String filter, String logGroup) {
-        try {
-            DeleteSubscriptionFilterRequest request = DeleteSubscriptionFilterRequest.builder()
-                    .filterName(filter)
-                    .logGroupName(logGroup)
-                    .build();
-
-            logs.deleteSubscriptionFilter(request);
-            System.out.printf("Successfully deleted CloudWatch logs subscription filter %s", filter);
-
-        } catch (CloudWatchException e) {
-            System.err.println(e.awsErrorDetails().errorMessage());
-            System.exit(1);
-        }
-    }
-}
-// snippet-end:[cloudwatch.java2.delete_subscription_filter.main]
