+++ javav2/example_code/cloudwatch/src/main/java/com/example/cloudwatch/events/PutRule.java
@@ -1,0 +1,78 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.example.cloudwatch.events;
+
+// snippet-start:[cloudwatch.java2.put_rule.main]
+// snippet-start:[cloudwatch.java2.put_rule.import]
+import software.amazon.awssdk.services.cloudwatch.model.CloudWatchException;
+import software.amazon.awssdk.services.cloudwatchevents.CloudWatchEventsClient;
+import software.amazon.awssdk.services.cloudwatchevents.model.PutRuleRequest;
+import software.amazon.awssdk.services.cloudwatchevents.model.PutRuleResponse;
+import software.amazon.awssdk.services.cloudwatchevents.model.RuleState;
+// snippet-end:[cloudwatch.java2.put_rule.import]
+
+/**
+ * Before running this Java V2 code example, set up your development
+ * environment, including your credentials.
+ *
+ * For more information, see the following documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+ */
+public class PutRule {
+    public static void main(String[] args) {
+        final String usage = """
+
+                Usage:
+                  <ruleName> roleArn>\s
+
+                Where:
+                  ruleName - A rule name (for example, myrule).
+                  roleArn - A role ARN value (for example, arn:aws:iam::xxxxxx047983:user/MyUser).
+                """;
+
+        if (args.length != 2) {
+            System.out.println(usage);
+            System.exit(1);
+        }
+
+        String ruleName = args[0];
+        String roleArn = args[1];
+        CloudWatchEventsClient cwe = CloudWatchEventsClient.builder()
+                .build();
+
+        putCWRule(cwe, ruleName, roleArn);
+        cwe.close();
+    }
+
+    /**
+     * Creates a new CloudWatch Events rule with the specified name and role ARN.
+     * The rule is configured to trigger every 5 minutes and is enabled upon creation.
+     *
+     * @param cwe The {@link CloudWatchEventsClient} used to interact with AWS CloudWatch Events.
+     * @param ruleName The name of the CloudWatch Events rule to be created.
+     * @param roleArn The ARN of the IAM role that CloudWatch Events can assume to execute the rule.
+     *
+     */
+    public static void putCWRule(CloudWatchEventsClient cwe, String ruleName, String roleArn) {
+        try {
+            PutRuleRequest request = PutRuleRequest.builder()
+                    .name(ruleName)
+                    .roleArn(roleArn)
+                    .scheduleExpression("rate(5 minutes)")
+                    .state(RuleState.ENABLED)
+                    .build();
+
+            PutRuleResponse response = cwe.putRule(request);
+            System.out.printf(
+                    "Successfully created CloudWatch events rule %s with arn %s",
+                    roleArn, response.ruleArn());
+
+        } catch (CloudWatchException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+}
+// snippet-end:[cloudwatch.java2.put_rule.main]
