+++ javav2/example_code/cloudwatch/src/main/java/com/example/cloudwatch/events/PutTargets.java
@@ -1,0 +1,82 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.example.cloudwatch.events;
+
+// snippet-start:[cloudwatch.java2.put_targets.main]
+// snippet-start:[cloudwatch.java2.put_targets.import]
+import software.amazon.awssdk.services.cloudwatch.model.CloudWatchException;
+import software.amazon.awssdk.services.cloudwatchevents.CloudWatchEventsClient;
+import software.amazon.awssdk.services.cloudwatchevents.model.PutTargetsRequest;
+import software.amazon.awssdk.services.cloudwatchevents.model.Target;
+// snippet-end:[cloudwatch.java2.put_targets.import]
+
+/**
+ * To run this Java V2 code example, ensure that you have setup your development
+ * environment, including your credentials.
+ *
+ * For information, see this documentation topic:
+ *
+ * https://docs.aws.amazon.com/sdk-for-java/latest/developer-guide/get-started.html
+ */
+public class PutTargets {
+    public static void main(String[] args) {
+        final String usage = """
+
+                Usage:
+                  <ruleName> <functionArn> <targetId>\s
+
+                Where:
+                  ruleName - A rule name (for example, myrule).
+                  functionArn - An AWS Lambda function ARN (for example, arn:aws:lambda:us-west-2:xxxxxx047983:function:lamda1).
+                  targetId - A target id value.
+                """;
+
+        if (args.length != 3) {
+            System.out.println(usage);
+            System.exit(1);
+        }
+
+        String ruleName = args[0];
+        String functionArn = args[1];
+        String targetId = args[2];
+        CloudWatchEventsClient cwe = CloudWatchEventsClient.builder()
+                .build();
+
+        putCWTargets(cwe, ruleName, functionArn, targetId);
+        cwe.close();
+    }
+
+    /**
+     * Adds a target to a specified CloudWatch Events rule.
+     *
+     * @param cwe The CloudWatchEventsClient used to interact with Amazon CloudWatch Events.
+     * @param ruleName The name of the CloudWatch Events rule to which the target will be added.
+     * @param functionArn The Amazon Resource Name (ARN) of the AWS Lambda function to be invoked by the target.
+     * @param targetId The unique identifier for the target.
+     *
+     */
+    public static void putCWTargets(CloudWatchEventsClient cwe, String ruleName, String functionArn, String targetId) {
+        try {
+            Target target = Target.builder()
+                    .arn(functionArn)
+                    .id(targetId)
+                    .build();
+
+            PutTargetsRequest request = PutTargetsRequest.builder()
+                    .targets(target)
+                    .rule(ruleName)
+                    .build();
+
+            cwe.putTargets(request);
+            System.out.printf(
+                    "Successfully created CloudWatch events target for rule %s",
+                    ruleName);
+
+        } catch (CloudWatchException e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+            System.exit(1);
+        }
+    }
+}
+// snippet-end:[cloudwatch.java2.put_targets.main]
